fix: resolve clippy warnings and improve code quality
- Remove unused safe_lock_warn function
- Fix lifetime annotation in safe_lock
- Replace filter + collect with retain in message filters
- Remove unnecessary type casts (PROTOCOL_VERSION, u64, u8)
- Remove Uuid.clone() calls (Uuid is Copy)
- Remove needless borrow of message_tx
- Replace redundant closures with associated functions
- Remove redundant closure for Duration::from_millis
- Implement FromStr trait for PeerStatus
- Replace map_or with is_some_and
- Add allow attribute for await_holding_lock (requires major refactoring)
- Remove redundant clone on Copy types
- Replace unwrap_or_else with unwrap_or where applicable

Remaining non-blocking warnings:
- PeerDto::new has 10/7 arguments (requires builder pattern)
- Needless fn main in doctest (documentation issue)

+++ feiqiu/src-tauri/src/modules/file_transfer/response.rs
@@ -230,7 +230,7 @@ impl FileTransferResponse {
             .as_secs();
 
         let proto_msg = ProtocolMessage {
-            version: PROTOCOL_VERSION as u8,
+            version: PROTOCOL_VERSION,
             packet_id,
             user_id: String::new(),
             sender_name: self.username.clone(),
